Move pattern utility functions to validation.ts
