+++ frontend/src/utils/index.ts
@@ -1,3 +1,3 @@
 export * from "./validation"
 export * from "./showToastOnError"
-export * from "./patterns"
+export * from "./date"
