+++ frontend/src/utils/validation.ts
@@ -1,3 +1,13 @@
+export const emailPattern = {
+  value: /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i,
+  message: "Invalid email address",
+}
+
+export const namePattern = {
+  value: /^[A-Za-z\s\u00C0-\u017F]{1,30}$/,
+  message: "Invalid name",
+}
+
 export const passwordRules = (isRequired = true) => {
   const rules: any = {
     minLength: {
